+++ src/app/auth/login/page.tsx
@@ -5,15 +5,18 @@ import LoginTerms from "@/components/auth/LoginTerms";
 import LoginWelcome from "@/components/auth/LoginWelcome";
 import { supabase } from "@/utils/database";
 import { Provider } from "@supabase/supabase-js";
+import { useSearchParams } from "next/navigation";
 
 export default function LoginPage() {
-
+	const searchParams = useSearchParams();
+	const isAppLogin = searchParams.get("app") === "true";
+	
 	const handleLogin = async (provider: Provider) => {
 		try {
 			const { error } = await supabase.auth.signInWithOAuth({
 			  provider,
 			  options: {
-				redirectTo: `${window.location.origin}/auth/callback`,
+				redirectTo: `${window.location.origin}/auth/callback/${isAppLogin ? "app" : "web"}`,
 			  },
 			});
 	  
@@ -21,7 +24,7 @@ export default function LoginPage() {
 			  throw error;
 			}
 		  } catch (error) {
-			console.error("Error loging in with Google:", error);
+			console.error("Error logging:", error);
 		  }
 	}
 
@@ -32,7 +35,7 @@ export default function LoginPage() {
 		>
 			<div className="h-fit w-full flex max-w-xl">
 				<div className="w-full h-full group p-6 rounded-xl border border-white/10 backdrop-blur-md bg-white/5 transition-all duration-300 shadow-lg">
-					<LoginWelcome />
+					<LoginWelcome isAppLogin={isAppLogin} />
 					<LoginOptions handleLogin={handleLogin}/>
 					<LoginTerms />
 				</div>
